Fix directory selection when BIOS is in an alternate folder
This will allow selecting a different folder on the first run instead
of having to place everything in /sdcard/dc first and then pick a new
folder to later move to. This will still notify the user of
requirements for having the BIOS in the /dc folder if they do not pick
an alternate location at this time.