Squash changes from andoidui into single commit Integrate previously locked settings with valid user warning
Put a note about the beta compile link in the README.md

Add a super fancy about that supports Github commit logs

Revise about with git log, Highlight current build in commits

Check for a null value when getting the current commit sha

Chinese Simplified Translation from @weihan1102

Rename Simplified Chinese res folder for proper complie

Add links to the website and beta compile site from about

Add option for custom mapping to set joystick to non-dpad

Fix a missing check for the user saved value of the A button

The Moga only operates one way, there is no variation test

Add region values, Set cable to match formatting of others

Reduce the number of random string array calls, jsCompat

It was unnecessary to keep getting the array, It was even less
necessary to keep getting the preference values. The port
identification letters are set multi-lingual values.

Label the about links to prevent any confusion (Not mirrors)

Missing a check on the motion events to determine compat

Add a check to verify Kitkat, Remove the standard override

The new API verification complains about the Override, so this is a
preventative step for if and when it is updated. It may also prevent
some issues with compatibility.

Run organize imports to reduce the amount of excess code

Add a controller option to go to custom layout from selection

Prevent a leak where multiple items were highlighted in list

@AndroidGX French Translation resources