Unlink custom from JS-Dpad, Rename appropriately
The preference name was a little misleading. The preference has the
same functionality, but was named to imply an inverse functionality. It
is now disabled by default, which may need to be reversed if the number
of games requiring the feature to split the two controls is outweighed
by the number for which the feature disables joystick functionality.