Permission aware file browsing, Restore menu size
Since the BIOS cannot be stored on external storage, there is
absolutely no reason to begin a search from that folder.

It makes more sense to start from a user-defined folder than the
default simply because an external card exists. A check has been added
to see if the preference is a default.