Increase the search depth slightly
This will account for storing games in a second-tier folder, such as
putting bins in one directory and chd in another with each game having
it’s own folder for bin and cue.
The BIOS search should handle 4 levels to account for the use of
/Android/data/com.reicast.emulator/files but should also ignore
directories, such as cache and obb, where the file will never be.