Fix hardware detection for Odroid-XU4 against newer kernels
On older kernels, the Odroid-XU4 would be identified as "ODROID-XU3". On newer kernels, such as the latest Ubuntu kernel for ODroid, the XU4 is now being identifed as "ODROID-XU4" which would cause the hardware detection to fail and have it build a generic ARM build instead of our desired "platform = odroidxu3".