send_mutex lock must be held when peeking socket
In absence of this change I noticed strange return values from 'recv()'. For example, 'len == -1' and 'errno == 0'. In some cases 'len == -1' would come back, and on the next attempt I would get 'len == 0'. I found it comes down to a need to synchronize when using the socket. 'refreshifNeeded()' has a similar issue which I had solved in the past by ensuring the lock is held. Not sure why it didn't occur to me as necessary here also, until now.