Fix bug involving refreshing multiple DreamLinks
`refreshIfNeeded` was only tearing down DreamLinkDevices for the single DreamLink identified as needing refresh. But `maple_ReconnectDevices()` applies to all the devices. If we call `maple_ReconnectDevices()` without first calling `tearDownDreamLinkDevices()` for all active DreamLinks, we get into an inconsistent state, where a subsequent call to `createDreamLinkDevices()` sees an existing device in the `dreamLinkVmus` list, and therefore doesn't update `MapleDevices`.

The solution I am going with, is to ensure that if *any* DreamLink needs devices refreshed, then `tearDownDreamLinkDevices()` is called for all the active DreamLinks, before `maple_ReconnectDevices()` is called.

As a "side benefit" of this, we can delete `dreamLinkNeedsRefresh`. In the path where we used to check that, we will always create devices for all the active DreamLinks now. There's no way to reduce work by trying to skip refreshing the ones whose configuration didn't change.

Also, `allDreamLinks` is renamed to `activeDreamLinks`, to indicate it stores only the active DreamLink for each port, and `DreamLink::refreshIfNeeded()` is now `DreamLink::needsRefresh()`, to reflect that while we do check for a changed device configuration, we don't actually change any of the MapleDevices.